+++ tests/test_type_float/test_type_float.py
@@ -72,6 +72,11 @@ def test_from_number():
     assert float.from_number(MyFloat2(18.87)) == 18
 
 
+def test_hex_fromhex():
+    s = (17.123).hex()
+    assert float.fromhex(s) == 17.123
+
+
 def test_all():
     test_float()
     test_inf()
@@ -79,6 +84,7 @@ def test_all():
     test_from_number()
     test_division()
     test_multiplication()
+    test_hex_fromhex()
 
 
 if __name__ == "__main__":